Update validation for more states, cut down extra features
