+++ src/css/components/form.pcss
@@ -37,10 +37,10 @@
       fill: inherit;
     }
   }
+}
 
-  &__message {
-    
-  }
+.cdr-input-messages {
+  
 }
 
 .cdr-input,
